bug(app-crash): fix app crash when no internet
The application crashes on the first instance it loads since there is no
internet connection.

This commit add onError method call on the observer that receive
data made on the endpoint of login in a user, without this method errors
are not being handled hence the crash. Add a toast to notify the user
of unavailable internet connection.

[Fixes 157103459]